removed usage of openAIKey for ollama
